feat: 호가 ui 작업: 추후 디자인 수정 필요
