+++ src/features/index.ts
@@ -1,3 +1,3 @@
 export { default as Trade } from "./Trades/ui/index";
-
-export {default as OrderBook} from "./OrderBook/ui/index";
+export { default as OrderBook } from "./OrderBook/ui/index";
+export { default as CandleChart } from "./CandleChart/ui/index";
